Organises dirs
