+++ src/components/Dashboard.tsx
@@ -11,10 +11,10 @@ import { visuallyHidden } from "@mui/utils";
 import { Link as RouterLink } from "react-router-dom";
 import { useCountryData } from "@/hooks/useCountryData";
 import { useTheme } from "@mui/material/styles";
-import formatPopulation from "@/functions/formatPopulation";
+import formatPopulation from "@/utils/functions/formatPopulation";
 import FilterMenu from "./FilterMenu";
 import { useState } from "react";
-import useStyles from "@/functions/useStyles";
+import useStyles from "@/utils/functions/useStyles";
 
 const Dashboard = () => {
   const theme = useTheme();
