+++ src/utils/functions/INTEGER_FORMATTER.ts
