+++ src/utils/functions/formatPopulation.ts
